# Express-Blueprint-main/apidocs.yml
openapi: 3.0.0
info:
  title: Express-Blueprint
  version: 1.0.0
servers:
  - url: http://localhost:8888
paths:
  /login:
    post:
      summary: Logs a user in
      description: Authenticates a user and returns a JWT token that can be used to access protected endpoints.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The username of the user to authenticate.
                password:
                  type: string
                  description: The password of the user to authenticate.
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    description: The ID of the authenticated user.
                  token:
                    type: string
                    description: The JWT token that can be used to access protected endpoints.
        '401':
          description: Invalid username or password
  /logout:
    post:
      summary: Logs a user out
      description: Invalidates the JWT token used to authenticate the user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Invalid token
  /register:
    post:
      summary: Registers a new user
      description: Creates a new user account with the specified username, email, and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                  description: The email address of the user to register.
                username:
                  type: string
                  description: The username of the user to register.
                password:
                  type: string
                  description: The password of the user to register.
              required:
                - mail
                - username
                - password
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Username or mail already in use
        '500':
          description: Internal server error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT